refactor(fs): now it is time to get rid of fs_loop and fs_loop_mutex
Here's the headline: when run in sync mode (last argument cb=NULL),
these functions don't actually use the uv_loop_t. However this is
not a documented guarantee.

The _vast_ majority of usecases of the functions run in the context
of the main loop, so we can just pass in `main_loop.uv`, just in case.
The only exceptions are:

- os_fopen() in init_startuptime() before main loop is initialized

We recently fixed fopen() to work in the windows build, so we
can just use fopen(). Amazing!

- nvim__get_runtime() has an implemention for worker threads.

This is not as bad as it seems, as it only uses the last available
cached path from the main thread. Thus, in only needs
os_file_is_readable() and os_is_dir() functionality. We can
create thread safe variants of these, which uses the luv_loop()
private to each worker thread.

Again the headline: The uv_fs module does not use the loop
parameter when running in sync mode. We could get rid of the extra
bookkeeping and just pass in loop=NULL. This works in the current
bundled version of libuv, but there is no guarntee this will keep
working in a future libuv update.