feat(api): add `delete` option for `nvim_buf_delete()`
Previously it was impossible to replicate the behavior of `:bdelete` with `nvim_buf_delete()`. The closest you could get would be to use `nvim_buf_delete()` with `unload = true` and then manually set `buflisted = false` for the buffer. But even then it would not be a perfect replication because `:bdelete` deletes some information for the buffer which unloading the buffer does not do. This PR adds a `delete` option which replicates the behavior of `:bdelete`.