fix(treesitter): clear callbacks when destroying language tree
Recursively removes all registered callbacks from a LanguageTree and its
children.  Without this cleanup it is possible for callbacks to
accumulate; for example if there is an autocommand to register a
callback when attaching to a buffer and the user executes `:edit` the
callback will be registered again.  With this change the user can
`destroy` the parser when it is detached from the buffer to release the
references to the callback functions.