fix(ui): delay win_viewport until the screen is updated
When there are pending updates to a Window, the win_viewport UI event is
delayed until the updates are sent. This ensures that there's no flush
between sending the viewport and updating of the lines corresponding to
the viewport.

Previously that happened, for example when scrolling up and down and
some of the lines were folded. But it could happen any time
w_viewport_invalid is set and ui_flush was called.