docs(Bram): A sad farewell
This is by no means up to the full community standards of NeoVim
documentation but I needed a starting point just to do some processing,
and I'm sure other people feel similar. I don't even feel comfortable
saying that much. I never met Bram Moolenaar. I never talked to him. I
was never active on the vim.org communities before NeoVim was created,
and was mostly on r/vim on reddit to interact with other Vim and NeoVim
people. But I used this tool he created almost every day since 2006, and
I always kept up to date with what he was doing with and for it.