vim-patch:b69b9d5e1753
Add filetype detection for eyaml files (vim/vim#12659)

https://github.com/voxpupuli/hiera-eyaml/ uses and produces the eyaml
format, which is simply yaml with some encrypted values.

It's convenient to edit the file without decrypting when not touching
encrypted values (or when you don't have access to the decryption key),
which is why vim should treat those files as yaml files.

https://github.com/vim/vim/commit/b69b9d5e175351c9a25507139954f6211f37fdd9

Co-authored-by: Max Gautier <mg@max.gautier.name>