fix(terminal): preserve $COLORTERM value from outer terminal
If the outer terminal claims to support truecolor through the $COLORTERM
environment variable, then it should not be modified by Nvim just
because 'termguicolors' is unset.

If an application running in :terminal writes a direct color sequence
(i.e. CSI 38 ; 2 ; <r> ; <g> ; <b> m) that sequence is forwarded
directly to the outer terminal emulator, where it is displayed in
truecolor if the terminal supports it, regardless of the value of
'termguicolors'. The only affect that 'termguicolors' has on colors
for applications in :terminal is how the 16 ANSI colors are displayed
(when 'termguicolors' is set, the ANSI colors are set with
g:terminal_color_x; otherwise, they are forwarded directly to the outer
terminal).

Ref: https://github.com/neovim/neovim/pull/13315

Fixes: https://github.com/neovim/neovim/issues/24717