feat(ui): forward :intro message properly to external UI:s
The intro message doesn't work very well with multigrid UI:s
The message is plainly drawn on the `default_grid` which is kinda ok
in TUI but doesn't provide any useful context for external UI:s.

Introduce a new `msg_intro` event with is used when `ext_multigrid`
and/or `ext_messages` is set.

Fixes #24705